requirements.txt
streamlit
openai
requests
python-dotenv
transformers
fpdf
serpapi
kaggle
PyMuPDF
